add fixes

+++ database/migrations/2025_05_15_170438_add_sent_at_to_posts_table.php
@@ -13,9 +13,11 @@ return new class extends Migration
     }
 
     public function down()
-    {
+{
+    if (Schema::hasColumn('posts', 'sent_at')) {
         Schema::table('posts', function (Blueprint $table) {
             $table->dropColumn('sent_at');
         });
     }
+}
 };
